controller: allow RBACSyncConfig to bind to ClusterRole
Issue:
https://github.com/cruise-automation/rbacsync/issues/12

Context:
There are valid use cases for binding a RBACSyncConfig to a ClusterRole.
Specifically this allows granting permission to namespaced resources defined in
the ClusterRole within the RoleBinding's namespace. This enables defining stanard
roles for the entire cluster.
https://kubernetes.io/docs/reference/access-authn-authz/rbac

Change:
The type check for role ref type was changed to allow ClusterRole. The docs were
updated to specify this is allowed and provide explanation for why.

Testing:
A new test was added with a purely made-up role ref type to test that invalid
role ref types do not work. Then an additional test was added for the new
ClusterRole role ref type check.

Signed-off-by: Seth Bunce <seth.bunce@getcruise.com>